pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dhis2eo"
version = "1.0.1"
description = "DHIS2 Earth Observation Toolkit"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "earthkit-data[netcdf,geotiff,geopandas,projection,cds]==0.16.*",
    "xarray>=2025.6,<2026",
    "rioxarray==0.19.*",
    "pandas>=2.3,<3",
    "geopandas>=1.1,<2",
    "numpy>=2.2,<3",
    "ecmwf-api-client>=1.6,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dhis2eo*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "ruff==0.14.*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (skipped by default)"
]
addopts = [
    "-m not integration"      # Skip tests marked as 'integration' by default
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # ignore unused imports in init files
